Requested changes + Refactor on servo_runtime
