+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod browser;
 pub mod cli;
 pub mod input;
+pub mod logger;
 pub mod output;
 pub mod platform;
